requirements.txt
aiofiles==24.1.0
aiohttp==3.10.3
asynciolimiter==1.0.0
cachetools==5.4.0
colorama==0.4.6
discord.py==2.3.2
openai==0.28.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32